dry up *Processors using process[T] framework
